requirements.txt
flask
flask_cors
gunicorn
pyjwt
werkzeug